Change statistics value data type to BigIntegerField.

+++ boostan/api/models.py
@@ -1,7 +1,9 @@
 from django.contrib.auth import get_user_model
 from django.contrib.humanize.templatetags.humanize import naturaltime
 from django.db import models
-from django.db.models import Max, Min, Sum
+from django.db.models import Max
+from django.db.models import Min
+from django.db.models import Sum
 from django.utils.translation import gettext as _
 from utils.general_model import GeneralModel
 
@@ -94,7 +96,7 @@ class Setting(models.Model):
 
 class Statistics(models.Model):
     name = models.CharField(verbose_name=_("Name"), max_length=256)
-    value = models.IntegerField(verbose_name=_("Value"), default=0)
+    value = models.BigIntegerField(verbose_name=_("Value"), default=0)
 
     class Meta:
         verbose_name = _("Statistics")
